+++ orders/models.py
@@ -2,5 +2,4 @@ from django.db import models
 
 # Create your models here.
 from .order_models.order import Order
-from .order_models.order_item import OrderItem
-from .order_models.order_item_feature import OrderItemFeature
+from .order_models.order_feature import OrderFeature
