+++ orders/order_models/order_feature.py
@@ -1,11 +1,11 @@
 from django.db import models
 
-from orders.order_models.order_item import OrderItem
+from orders.order_models.order import Order
 from services.service_models.feature import Feature
 
 
-class OrderItemFeature(models.Model):
-    order_item = models.ForeignKey(OrderItem, on_delete=models.CASCADE, related_name="order_item_features")
+class OrderFeature(models.Model):
+    order = models.ForeignKey(Order, on_delete=models.CASCADE, related_name="order_features")
     feature = models.ForeignKey(Feature, on_delete=models.CASCADE)
     extra_cost = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
